refactor(cli): divide all commands
- Remove experimental warning prevention
- Split all commands into separate files
